subcoin_rpc_bitcoind

Trait MiningApiServer

pub trait MiningApiServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required method
    fn estimate_smart_fee(
        &self,
        conf_target: u32,
        estimate_mode: Option<String>,
    ) -> Result<EstimateSmartFee, Error>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the MiningApi RPC API.

Required Methods

fn estimate_smart_fee( &self, conf_target: u32, estimate_mode: Option<String>, ) -> Result<EstimateSmartFee, Error>

Estimates the approximate fee per kilobyte needed for a transaction to begin confirmation within conf_target blocks.

Returns estimate fee rate in BTC/kvB.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.